+++ audiobookshelf_config.example.lua
@@ -1,4 +1,5 @@
+-- plugins/audiobookshelf.koplugin/audiobookshelf_config.lua
 return {
-    token = 'your api key here',
-    server = 'your audiobookshelf instance url here'
+    ["token"] = 'your api key here',
+    ["server"] = 'your audiobookshelf instance url here'
 }