+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.1 (2025-04-25)
+
+Add search functionality
+
+Allow user to update settings from plugin
+
+## 0.1.0 (2025-04-14)
+
+First functionality to allow user to download books
+
 ## 0.0.1 (2025-04-10)
 
 Initial release